requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.2.0
PyYAML>=6.0
pillow>=10.0.0
tqdm>=4.66.0

# Z-Image-Turbo (Diffusers)
git+https://github.com/huggingface/diffusers
transformers>=4.40.0
accelerate>=0.30.0
peft>=0.11.0
safetensors>=0.4.0
